requirements.txt
Django==4.2.13
djangorestframework==3.15.1
requests==2.32.3